Fix: Incorrect directory path if tool is installed by package
Currently, if we install d1-graphics-tool with a .deb package, it will try
to resolve the path to save a .json file to /usr/bin, since currently we
save the configuration file wherever the binary is. User doesn't have an
access to write to /usr/bin/ directly and this ends up with a problem in
reading/writing to a config, resetting the config everytime app is being
used.

This patch changes it, and leaves the configuration for the file to be
saved in the application directory if it is writable, otherwise in the
configuration folder:

Windows: "C:/Users/<USER>/AppData/Local/D1GraphicsTool/"
MacOS: "~/Library/Preferences/D1GraphicsTool/"
Linux: "/home/user/.config/D1GraphicsTool/"
Android: "<APPROOT>/files/settings/"
iOS: "<APPROOT>/Library/Preferences/D1GraphicsTool"